legaldoc/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
scikit-learn>=1.1.0
pandas>=1.4.0
numpy>=1.21.0
opencv-python>=4.6.0
pytesseract>=0.3.9
PyPDF2>=2.10.0
pdf2image>=1.16.0
pillow>=9.2.0
fastapi>=0.78.0
uvicorn>=0.18.0
pydantic>=1.9.0
pyyaml>=6.0
pytest>=7.1.0
easyocr>=1.6.0
spacy>=3.4.0
nltk>=3.7
xgboost>=1.6.0
lightgbm>=3.3.0
